pyproject.toml
[project]
name = "krishnaik-simple-rag-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.5",
    "datasets>=2.14.0",
    "faiss-cpu==1.12.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "jq>=1.10.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.80",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langchain-text-splitters>=0.3.11",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "notebook>=7.5.0",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "opik>=1.9.40",
    "pandas>=2.3.3",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "ragas==0.3.7",
    "seaborn>=0.12.0",
    "sentence-transformers>=0.2.5.1",
    "tiktoken>=0.12.0",
]
